feat: add a second figure to test appearance of scaling
One of the problems with some approaches to SVG scaling is that
different figures are scaled by different amounts: for example,
specifying a width of 500px will scale intrinsically small figures
more than intrinsically large ones. This PR adds two more figures to
the sample slide deck to check this. It appears that figures are
scaled by the same amount (so that, for example, embedded text appears
the same size in different figures after scaling), but the concept map
figure appears to be offset to the right: is it being centered and
*then* being scaled?